+++ backend/src/main/java/codezap/auth/dto/LoginAndCredentialDto.java
@@ -1,0 +1,12 @@
+package codezap.auth.dto;
+
+import codezap.auth.dto.response.LoginResponse;
+
+public record LoginAndCredentialDto(
+        LoginResponse loginResponse,
+        String credential
+) {
+    public static LoginAndCredentialDto from(LoginResponse loginResponse, String credential) {
+        return new LoginAndCredentialDto(loginResponse, credential);
+    }
+}
