+++ backend/src/main/java/codezap/auth/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package codezap.member.dto.request;
+package codezap.auth.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
