+++ backend/src/main/java/codezap/auth/dto/response/LoginResponse.java
@@ -1,0 +1,12 @@
+package codezap.auth.dto.response;
+
+import codezap.member.domain.Member;
+
+public record LoginResponse(
+        long memberId,
+        String username
+) {
+    public static LoginResponse from(Member member) {
+        return new LoginResponse(member.getId(), member.getUsername());
+    }
+}
