--- backend/src/main/java/codezap/member/dto/response/LoginResponse.java
@@ -1,7 +1,0 @@
-package codezap.member.dto.response;
-
-public record LoginResponse(
-        long memberId,
-        String username
-) {
-}
